push after pull requesst

+++ Fashion-Store/src/main/java/Fashion_Store/Fashion_Store/Model/Country.java
@@ -9,7 +9,7 @@ import jakarta.persistence.*;
 public class Country {
     @Id
     @GeneratedValue(strategy = jakarta.persistence.GenerationType.IDENTITY)
-    private long countryId;
+    private Integer countryId;
     @Column(nullable = false,unique = true)
 
     private String countryName;
