+++ Fashion-Store/src/main/java/Fashion_Store/Fashion_Store/Model/Supplier.java
@@ -20,7 +20,7 @@ import jakarta.persistence.Table;
 public class Supplier {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long supplierId;
+    private Integer supplierId;
     @Column(nullable = false)
     private String supplierName;
